completed the agent commission negotiation flow

+++ src/java/com/tp/neo/controller/helpers/SaleItemObject.java
@@ -29,7 +29,7 @@ public class SaleItemObject {
     public float monthlyPayPerUnit;
     public float monthlyPayPerQuantity;
     public float productMinimumMonthlyPayment;
-    public float commp ;
+    public Double commp ;
